[#113] refactor: 모달 컴포넌트 분리
